+++ docs/src/routes/layers.tsx
@@ -1,6 +1,6 @@
 import {Markdown} from '~/components/markdown';
 import ref from '../reference/latest';
-import {Items} from '../components/items';
+import {Items} from '~/components/Items/items';
 import {AppropriateImage} from '~/components/appropriate-image/appropriate-image';
 import Caption from '~/components/caption';
 import combineItems from '~/util/combine-items';
